Group testcases from test_server_routes.py into their appropriate files
