+++ wikilabels/tests/test_form_builder_routes.py
@@ -1,0 +1,5 @@
+from .routes_fixture import app  # noqa
+
+
+def test_form_builder(client):
+    assert client.get("/form_builder/")._status_code == 200
