+++ wikilabels/tests/test_gadget_routes.py
@@ -1,0 +1,5 @@
+from .routes_fixture import app  # noqa
+
+
+def test_gadget(client):
+    assert client.get("/gadget/")._status_code == 200
