+++ wikilabels/tests/test_root_routes.py
@@ -1,0 +1,5 @@
+from .routes_fixture import app  # noqa
+
+
+def test_root(client):
+    assert client.get("/")._status_code == 200
