+++ wikilabels/tests/test_stats_routes.py
@@ -1,0 +1,5 @@
+from .routes_fixture import app  # noqa
+
+
+def test_stats(client):
+    assert client.get("/stats/")._status_code == 200
