+++ wikilabels/tests/test_ui_routes.py
@@ -1,0 +1,5 @@
+from .routes_fixture import app  # noqa
+
+
+def test_ui(client):
+    assert client.get("/ui/")._status_code == 200
