+++ wikilabels/tests/test_versions_routes.py
@@ -1,0 +1,5 @@
+from .routes_fixture import app  # noqa
+
+
+def test_versions(client):
+    assert client.get("/versions/")._status_code == 200
